[core] Convert native image loader to float

+++ Native/libs/bve-core/include/core/image/loader.hpp
@@ -8,12 +8,12 @@ namespace bve::core::image {
 	  public:
 		explicit Loader(std::string filename);
 		glm::ivec2 dimensions() const noexcept;
-		const uint8_t* data() const noexcept;
+		const float* data() const noexcept;
 		bool valid() const noexcept;
 		~Loader();
 
 	  private:
 		glm::ivec2 dimensions_;
-		uint8_t* data_;
+		float* data_;
 	};
 } // namespace bve::core::image
